Move all task related query handling to new package duke.query.task
All query handling are in duke.query.

Moving all module specific query handling to its own package will
unclutter the query package.
