+++ src/main/java/duke/query/InvalidCommandParamException.java
@@ -3,7 +3,7 @@ package duke.query;
 import duke.DukeException;
 
 public class InvalidCommandParamException extends DukeException {
-    InvalidCommandParamException(String msg) {
+    public InvalidCommandParamException(String msg) {
         super(msg);
     }
 }
