+++ src/main/java/duke/query/task/EventQueryHandler.java
@@ -1,10 +1,12 @@
-package duke.query;
+package duke.query.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 import duke.DukeException;
+import duke.query.InvalidCommandParamException;
+import duke.query.Query;
 import duke.task.Task;
 import duke.task.TaskTracker;
